# crypto: optimize wasm bindings