Makes `std/crypto` hashes faster than `std/hash`. Faster than native WebCrypto for small buffers.

```
benchmark SHA-256                64B runtime WebCrypto (target) ...
    10 runs avg: 0.14642090000015742ms
benchmark SHA-256                64B std@0.102.0/hash WASM (baseline) ...
    10 runs avg: 0.13566670000054729ms
benchmark SHA-256                64B std/crypto WASM   (you are here) ...
    10 runs avg: 0.02513749999998254ms
...
benchmark SHA-512           4194304B runtime WebCrypto (target) ...
    10 runs avg: 7.249925099999382ms
benchmark SHA-512           4194304B std@0.102.0/hash WASM (baseline) ...
    10 runs avg: 7.745979000000807ms
benchmark SHA-512           4194304B std/crypto WASM   (you are here) ...
    10 runs avg: 7.6258332999997945ms
benchmark SHA-512          67108864B runtime WebCrypto (target) ...
    10 runs avg: 113.68318340000042ms
benchmark SHA-512          67108864B std@0.102.0/hash WASM (baseline) ...
    10 runs avg: 123.60374999999912ms
benchmark SHA-512          67108864B std/crypto WASM   (you are here) ...
    10 runs avg: 120.48269170000057ms
benchmark SHA-512         524291328B runtime WebCrypto (target) ...
    10 runs avg: 901.8525874000006ms
benchmark SHA-512         524291328B std@0.102.0/hash WASM (baseline) ...
    10 runs avg: 965.9354250000018ms
benchmark SHA-512         524291328B std/crypto WASM   (you are here) ...
    10 runs avg: 910.8596960000025ms
```

- remove dependency on wasm_bindgen. (build CLI and bindings)
- non-fallible rust code.
- no special object heap, hence no external imports. avoids re-entering JS/WASM multiple times.