@iuioiua 2022-11-07
@littledivy, do these optimisations still have the potential to apply to the `std/crypto` of today?