@littledivy 2022-11-23
This PR is quite old now. This might be worth looking into later on as the improvement is only for small buffers